.history/requirements_20230704064626.txt
cloudinary==1.33.0
firebase_admin==6.1.0
nltk==3.8.1
openai
Pillow
protobuf==4.23.3
python-slugify==8.0.1
Requests==2.31.0
spacy==3.5.3
stability_sdk
streamlit==1.23.1
textblob==0.17.1
langchain
python-dotenv
